requirements.txt
pillow
pyexiftool